feat: allows probes to return single result value
Now probs can retun values which get saved in test artifacts. The name of artifcat name is the same as probe function name used.
For example if probe exited due timeout it can return False, so a test relyin on it can use this value instead of executing the same tests one more time